apps/production/Docker/requirements.txt
dash
plotly-express
pandas
dash_bootstrap_components